feat(*): 修复水平越权攻击

+++ shop-user-service/src/main/java/net/shop/service/impl/AddressServiceImpl.java
@@ -27,7 +27,10 @@ public class AddressServiceImpl implements AddressService {
 
     @Override
     public AddressVO detail(Long id) {
-        AddressDO addressDO = addressMapper.selectOne(new QueryWrapper<AddressDO>().eq("id", id));
+
+        LoginUser loginUser = LoginInterceptor.threadLocal.get();
+
+        AddressDO addressDO = addressMapper.selectOne(new QueryWrapper<AddressDO>().eq("id", id).eq("user_id",loginUser.getId()));
         if(addressDO==null){
             return null;
         }
@@ -74,7 +77,8 @@ public class AddressServiceImpl implements AddressService {
      */
     @Override
     public int delete(int addressId) {
-        int rows = addressMapper.delete(new QueryWrapper<AddressDO>().eq("id", addressId));
+        LoginUser loginUser = LoginInterceptor.threadLocal.get();
+        int rows = addressMapper.delete(new QueryWrapper<AddressDO>().eq("id", addressId).eq("user_id",loginUser.getId()));
         return rows;
     }
 
@@ -89,7 +93,7 @@ public class AddressServiceImpl implements AddressService {
 
         List<AddressDO> list = addressMapper.selectList(new QueryWrapper<AddressDO>().eq("user_id", loginUser.getId()));
 
-        //lamda表达式,对象流式转换
+        //lambda表达式,对象流式转换
         List<AddressVO> addressVOList = list.stream().map(obj -> {
             AddressVO addressVO = new AddressVO();
             BeanUtils.copyProperties(obj, addressVO);
